+++ lib/FibClock/FibClock.h
@@ -66,6 +66,74 @@ const Color fallMap[4] = {
 };
 // -----------------------------------------------------------------------------
 
+// ------ Holiday Color Sets ---------------------------------------------------
+#define NEWYEAR 101
+const Color newyearMap[4] = {
+    {red : 255, green : 255, blue : 255}, // white
+    {red : 255, green : 255, blue : 0},   // yellow
+    {red : 0, green : 150, blue : 200},   // blue
+    {red : 175, green : 0, blue : 255},   // purple
+};
+
+#define VALENTINES 214
+const Color vdayMap[4] = {
+    off,
+    {red : 255, green : 0, blue : 0},     // red
+    {red : 255, green : 255, blue : 255}, // white
+    {red : 255, green : 64, blue : 64},   // pink
+};
+
+#define BIRTHDAY 219
+const Color birthdayMap[4] = {
+    {red : 0, green : 255, blue : 0},   // green
+    {red : 255, green : 0, blue : 255}, // purple
+    {red : 0, green : 0, blue : 255},   // blue
+    {red : 255, green : 255, blue : 0}, // yellow
+};
+
+#define STPATTY 317
+const Color stpattyMap[4] = {
+    off,
+    {red : 0, green : 255, blue : 0},     // green
+    {red : 255, green : 255, blue : 255}, // white
+    {red : 40, green : 255, blue : 40},   // light-green
+};
+
+#define INDYPENDY 704
+const Color indyMap[4] = {
+    off,
+    {red : 255, green : 0, blue : 0},     // red
+    {red : 255, green : 255, blue : 255}, // white
+    {red : 0, green : 0, blue : 255},     // blue
+};
+
+#define HALLOWEEN 1031
+const Color halloweenMap[4] = {
+    {red : 255, green : 255, blue : 255}, // white
+    {red : 255, green : 100, blue : 0},   // orange
+    {red : 255, green : 0, blue : 255},   // purple
+    {red : 0, green : 255, blue : 0},     // green
+};
+
+// Not *exactly* the correct day, but close enough :)
+#define THANKSGIVING 1125
+const Color thanksgivingMap[4] = {
+    {red : 255, green : 255, blue : 255}, // white
+    {red : 255, green : 0, blue : 0},     // red
+    {red : 240, green : 255, blue : 0},   // yellow
+    {red : 255, green : 100, blue : 0},   // orange
+};
+
+// red, green, white, light-blue
+#define CHRISTMAS 1225
+const Color christmasMap[4] = {
+    {red : 255, green : 255, blue : 255}, // white
+    {red : 0, green : 255, blue : 0},     // green
+    {red : 255, green : 0, blue : 0},     // red
+    {red : 75, green : 150, blue : 255},  // light-blue
+};
+// -----------------------------------------------------------------------------
+
 // Spring: 3-20  to 6-20   => 320 to 620
 // Summer: 6-21  to 9-21   => 621 to 921
 // Fall:   9-22  to 12-20  => 922 to 1220